pyproject.toml
[project]
name = "data-insight-tool"
version = "0.1.0"
description = "Data insight tool with PostgreSQL MCP server"
requires-python = ">=3.10.6"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "fastmcp>=0.1.0",
    "python-dotenv>=1.1.1",
    "nest-asyncio>=1.6.0",
]
